Updates to CL fieldnames and error filling in fill nans.
merge_field_names: updated the SEAC4RS and DC3 sections to return
the correct field for chemiluminescence NO2.

fill_nans: significant change, added the ability to accept a vector
of error values and interpolate them to fill nan, not using linear
interpolation but the dependence of error on linear interpolation.