Misc. updates to util functions
Added titlef.m - a function that automatically parses
the string passed to title using sprintf. This allows
line breaks and other special characters like tabs.

merge_field_names - added INTEX-B data, which required
the addition of PRESSURE and TEMPERATURE fields.  Also
moved the choice of range files in here so it doesn't
need to be copied in different functions.

lsqfitgm - No change to code (except an extra output that
is no longer needed); this just includes a comment describing
the relationship between the std. dev. of the slope and the std.
error described in linear regression texts.

fill_nans - fixed a bug with the interaction of clipping and the
error vector.