Various updates to undiscovered functions, Matlab 2014b functionality. MODIFIED FILES:     merge_field_names: given the ability to return ground site info (field names and directory) and ground site info added for DISCOVER MD and CA campaigns.
    containedin: Matlab's builtin "ismember" function does what
contained in does, so it is now deprecated. For now, it will
issue a warning to that effect when used.

    make_empty_struct_from_cell: replaced containedin with
ismember. Minor edits to comments for clarity.

    tilefigs: edited to handle how Matlab 2014b changed figure
handles - they are now graphics objects rather than scalar numbers.
Consequently to order the figures by figure number needed code to
extract the figure number from the handles.