New error type added and the ability to insert sprintf formatting in custom errors.  Specifics:
filenotfound message changed to '<file> does not exist'

New error type: notimplemented.  Used for cases where a feature is
planned (and so the possibility of reaching it exists) but is not
implemented yet.

callError now passes the messages through sprintf. This allows both
the use of formatting strings like %s or %02d but requires certain
characters (%, \) to be escaped.

callCustomError now uses sprintf as well. It checks that the number of
format strings matches the number of additional inputs when called.

NEW FILE:
    .gitattributes - specifies that .mat, .fig, .jpg, and .png, and
.pdf files should be treated as binary.

r