Again an accumulated batch of updates.
MODIFIED FILES:
    state_outlines.m - modified so that it doesn't open a figure when returning
state border lat/lon if no figure open.

NEW FILES:
    calc_fit_line.m, plot_fit_line.m - moved over from the AircraftProfiles.git
repo. Updated from the last commit in that repo so that you can vary how the
x-coordinate is set (based on data or plot limits).

    angle_diffd.m - function for calculating the difference in angles (in
degrees) that accounts for the fact that 0 and 360 are the same, and that angles
near those may have smaller actual differences than their values suggest.

    make_filename_safe.m - makes strings path safe by removing most special
characters.  This includes slashes, as it is intended for filenames rather than
full paths.

    ndens_air.m - compute the number density of air given a temperature and
pressure.

    save_fig_slices.m - saves individual 2D slices of a 3D array as separate
figures (good for stitching together into an animation).