A collection of new functions and fixes made while responding to reviewers for Laughner & Cohen, AMT, 2017.
MODIFIED FILES:
    JLLErrors.m - badgeo error method can now take additional arguments for
sprintf-like insertion

    iscellcontents.m - updated an old syntax for JLLErrors where its methods
didn't call the error themselves. Also fixed an error called when the test
function returns a non-scalar value; simplified the error message since the test
function isn't always a string.

NEW FILES:
    clipmat.m - convenience function to restrict an array's values to a
specified range.

    dirff.m - wrapper around DIR() that automatically prepends the search path
to each file's name.

    files_input.m - similar in concept to VALIDATE_DATE(), this function takes
an input intended to be a list of files (which can be one of three formats),
checks that the input is one of those formats, and converts to a cell array. Can
do other checking as well.

    inpolygon_mg.m - a wrapper around INPOLYGON() for cases where the
coordinates make up a mesh grid, this can cut down the arrays more efficiently
since it knows it only has to look at one row or column of the coordinate
arrays.

    randrange.m - a convenience function that generates an array of random
floating point numbers between a given range of values.