4 new utilities and 3 changes useful in the BEHR v3 analysis
MODIFIED FILES:
    combine_plots.m - when specifying dimensions, one can be given as 0
to indicate that it should be calculated to fit the number of plots needed
within the constraint of the other dimension. Also, added option to auto-
matically rescale the combined plot figure to keep all of the subfigures
roughly their original size, or some fraction of it.

    mat2latex.m - if using asymmetrical uncertainty, allow the user to give
one uncertainty as a NaN indicating that this uncertainty is actually
symmetrical. This way asymmetrical and symmetrical uncertainties can coexist
in the same table.

    update_params.m - added ability to remove, as well as add or modify,
parameters.

NEW FILES:
    RunningAverage.m - a class that can be used to do running averages. Once
instantiated, data (with or without weights) can be added to it and it will
keep a running sum of the weighted data and the weights. You may calculate
the current weighted average at any time.

    histstats.m - plots a histogram with the mean marked as a red dashed line
and the mean, 1 std. deviation, median, and upper and lower quartiles printed
directly on the plot.

    nanadd.m - convenience function to apply nansum2 to multiple arrays more
easily. Normally C = A + B will cause a NaN to appear in C anywhere A or B has
a NaN, but often we want NaNs treated as zeros. C = nanadd(A, B) will do that,
putting a NaN in C only if both A and B have a NaN as that element.

    ncdf_var_by_name.m - returns the variable schema for a variable of a given
name in a netCDF file schema.