Collection of changes supporting BEHR v3 validation
MODIFIED FILES:
    calc_fit_line.m - new option to add an asterisk to the R2 value if the slope
is statistically different from 0.

    state_outlies.m - now keeps x and y limits that were in place before it was
called.

    udpate_params.m - for 'remove' mode, can give a structure as the third
argument and it will remove parameters matching the field names of the
structure.

NEW FILES:
    cat_cell_contents.m - concatenates different size numeric arrays from a cell
array with padding.

    difftol.m - check if each element of two arrays are within machine precision
of each other.

    opt_ask_multichoice.m - wrapper around ask_multichoice that assumes the
value returned by ask_multichoice may be optionally given to a function. This
takes that value, checks if it is a "blank" value and asks the user to choose a
value if it is. If not, then it checks if the given value is one of the allowed
choices.

    opt_ask_yn.m - a similar wrapper around ask_yn.

    slope_significant.m - test if a slope if statistically significant. By
default, tested to be different from 0 with 95% confidence, but both of those
parameters can be changed.

    slope_significant_difference.m - test if two slopes are statistically
different.