Collection of updates supporting BEHR v3.0 validation
MODIFIED FILES:
    advInputParser.m - no longer subclassed from inputParser, handles all
parsing itself, necessary rewrite to properly handle parameters mixed with
optional arguments that consume character inputs.

    bar_errors.m - modified to allow asymmetrical error bars on the plots.

    calc_fit_line.m - includes number of points used in the returned LineData
structure.

    calc_plot_limits.m - modified to use advInputParser instead of manually
parsing flag options. Added ability to restrict limits to some number of
standard deviations and corrected a bug in checking the limits when one side is
to be forced to 0.

    combine_plots.m - added ability to scale combined figure by different
amounts on the horizontal and vertical axes.

    difftol.m - modified to only do the tolerance tests on non-NaN values and
rely on the NaN XOR later to catch improper NaNs, fixes a bug where the test
failed for NaN values.