Range of updates, mainly in support of the lifetime analysis
MODIFIED FILES:
    ask_multiselect.m - removed unused 'rangeselect' parameter

    ask_number.m - properly print the default value if it is a range

    calc_fit_line.m - added ID to the "NaNs detected" warning so that it can be
suppressed; also, the intercept is set to 0 for the fit through the origin,
rather than NaN, since it is actually 0.

    copy_structure_fields.m - fixed a bug that occurred when using the 'missing'
option, if none were missing, it would try to copy all of the fields in "dest".
The proper behavior is to not copy anything.

    label_subfigs.m - replaced ad hoc input parsing with advInputParser +
update_params.

    make_latex_table.m - added options for "environment" (to change from a
standard table to, e.g. sidewaystable) and "center" (to automatically add a
\centering command)

    plot_changes.m - added option for connector format.

    plot_error_envelope_x.m - updated from old Matlab syntax (addParamValue ->
addParameter and got away from using figure numbers) and made it able to handle
non-vector inputs.

    plot_fit_line.m - added 'orth-origin' to the help text and got rid of the
in-parser test that 'regression' is one of the allowed ones; will just rely on
calc_fit_line for that.

    slope_significant.m - calculates the number of values as the number of
non-NaN points, rather than the number of all points, since NaNs don't actually
contribute information to the slope.

    update_params.m - added "missing" option, which will only append missing
parameters, not overwrite existing ones.